Re-factor plugin based on advice from Gradle engineer:
The solution I had before didn't follow the gradle way.
After discussions on the Gradle Forum between myself and
one of their engineers, he recommended using the withId
method.  This allows to to configure a plugin without
applying it.  When the plugin is applied, along with this
custom plugin, then the configuration contained within the
'withId' block is used, else it is ignored.  Each library
or application that applies this Plugin, should also apply
the Java or War plugin.